// frontend/shared/src/api/model/generated/assetUserRole.ts
/**
 * Generated by orval v6.24.0 🍺
 * Do not edit manually.
 * Navtrack.Api
 * OpenAPI spec version: 1.0.0
 */

export type AssetUserRole = (typeof AssetUserRole)[keyof typeof AssetUserRole];

// eslint-disable-next-line @typescript-eslint/no-redeclare
export const AssetUserRole = {
  Owner: "Owner",
  Viewer: "Viewer"
} as const;