// frontend/shared/src/api/model/generated/organizationUserRole.ts
/**
 * Generated by orval v6.31.0 🍺
 * Do not edit manually.
 * Navtrack.Api
 * OpenAPI spec version: 1.0.0
 */

export type OrganizationUserRole =
  (typeof OrganizationUserRole)[keyof typeof OrganizationUserRole];

// eslint-disable-next-line @typescript-eslint/no-redeclare
export const OrganizationUserRole = {
  Owner: "Owner",
  Member: "Member"
} as const;